requirements.txt
cffi==1.15.1
click==8.1.3
cryptography==39.0.0
dnspython==2.3.0
Flask==2.2.2
Flask-PyMongo==2.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pycparser==2.21
PyJWT==2.6.0
pymongo==4.3.3
Werkzeug==2.2.2
